backend-python-v2/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
web3
pydantic-settings
pydantic